nixos/mautrix-discord: refactor
After fixing the registration bug I determined that the module was way too verbose and did not need to provide all of the options provided in the application. I've stopped explicitly declaring all options that aren't needed to make a basic version of the module function. The assertions have also been updated to mirror this fact and will alert a user when the specific required options are unset. A new documnetation file was also added as this module, like other mautrix modules, is quite complex and may need more explaining.